+++ src/components/modal/modal.js
@@ -1031,20 +1031,40 @@ Modal.prototype = {
         return;
       }
 
+      self.setFocusableElems();
+      const focusableElements = $(self.focusableElems).not('.modal-header .searchfield');
+      const hiddenFieldTagNames = ['INPUT', 'BUTTON'];
+
       // When changes happen within the subtree on the Modal, rebuilds the internal hash of
       // tabbable elements used for retaining focus.
-      self.changeObserver = new MutationObserver(() => {
-        self.setFocusableElems();
+      self.changeObserver = new MutationObserver((mutationsList) => {
+        let updateFocusableElems = false;
+
+        mutationsList.forEach((mutation) => {
+          if (['childList', 'subtree'].includes(mutation.type)) {
+            updateFocusableElems = true;
+          } else {
+            const mutationTarget = $(mutation.target);
+            if (hiddenFieldTagNames.includes(mutation.target.tagName)) {
+              if (mutationTarget.is(':visible')) {
+                mutationTarget.attr('tabindex', '0');
+              } else {
+                mutationTarget.attr('tabindex', '-1');
+              }
+            }
+          }
+        });
+
+        if (updateFocusableElems) {
+          self.setFocusableElems();
+        }
       });
-      self.changeObserver.observe(self.element[0], { childList: true, subtree: true });
-      self.setFocusableElems();
-
-      const focusableElements = $(self.focusableElems).not('.modal-header .searchfield');
-
-      // The element/s will be disabled if detects that it has inline display: none; style.
-      if (focusableElements.css('display') === 'none') {
-        focusableElements.not(':visible').attr('disabled', 'disabled');
-      }
+      self.changeObserver.observe(self.element[0], {
+        attributes: true,
+        attributeFilter: ['class', 'style'],
+        childList: true,
+        subtree: true,
+      });
 
       let focusElem = focusableElements.not(':hidden').first();
 
@@ -1313,6 +1333,13 @@ Modal.prototype = {
     const ignoredSelectors = [
       'option'
     ];
+
+    // Pre-configure some element types that can be visually-hidden when the Modal opens
+    // to make it impossible to become focused while hidden (see infor-design/enterprise#6806)
+    const nonFocusableElems = $(this.element).find('input:hidden');
+    nonFocusableElems.each((i, elem) => {
+      elem.tabIndex = -1;
+    });
 
     const elems = DOM.focusableElems(this.element[0], extraSelectors, ignoredSelectors);
     this.focusableElems = elems;