+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ## v4.60.0 Fixes
 
+- `[About]` Added event exposure in about component. ([NG#1124](https://github.com/infor-design/enterprise-ng/issues/1124))
 - `[Actionsheet]` Fixed an Angular issue where the `renderRootElems` method was not re-rendered when going to other action sheet test pages due to SPA routing concept. ([NG#1188](https://github.com/infor-design/enterprise-ng/issues/1188))
 - `[Calendar]` Fixed an issue where you could not have more than one in the same page. ([#6042](https://github.com/infor-design/enterprise/issues/6042))
 - `[Column]` Fix a bug where bar size is still showing even the value is zero in column chart. ([#5911](https://github.com/infor-design/enterprise/issues/5911))
@@ -40,6 +41,7 @@
 - `[Icon]` Fixed the translate icon so it can take a color, fixed the tag icon as it was rendered oddly. ([#5870](https://github.com/infor-design/enterprise/issues/5870))
 - `[Listbuilder]` Fix on disable bug: Will not enable on call to enable() after disable() twice. ([#5885](https://github.com/infor-design/enterprise/issues/5885))
 - `[Locale]` Changed the text from Insert Anchor to Insert Hyperlink. Some translations my still reference anchor until updated from the translation team. ([#5987](https://github.com/infor-design/enterprise/issues/5987))
+- `[Modal]` Fixed a bug on hidden elements not focusable when it is turned visible. ([#6086](https://github.com/infor-design/enterprise/issues/6086))
 - `[Modal]` Fixed a regression bug where elements inside of the tab panel were being disabled when its `li` tab is not selected (is-selected class) initially. ([NG#1210](https://github.com/infor-design/enterprise-ng/issues/1210))
 - `[Searchfield]` Fixed UI alignment of close icon button (searchfield) in Datagrid. ([#5954](https://github.com/infor-design/enterprise/issues/5954))
 - `[Tabs Module]` Fixed UI alignment of close icon button on mobile view([#5951](https://github.com/infor-design/enterprise/issues/5951))